+++ __tests__/Base.spec.ts
@@ -79,6 +79,19 @@ describe(Base, () => {
           expect(requestError.error).toBe('Not found')
         }
       })
+
+      it('throws an ErrorObject', async () => {
+        const promise = model.withRequest('fetching', Promise.reject({ foo: 'bar' }))
+
+        try {
+          await promise
+        } catch (requestError) {
+          expect(requestError).toBeInstanceOf(ErrorObject)
+          expect(requestError.payload).toEqual({ foo: 'bar' })
+          expect(requestError.error).toEqual(null)
+          expect(requestError.requestResponse).toEqual(null)
+        }
+      })
     })
   })
 
@@ -117,16 +130,14 @@ describe(Base, () => {
     })
   })
 
-  describe('rpc(label, endpoint, options)', () => {
+  describe('rpc(endpoint, options, label)', () => {
     let request
     let spy
 
     beforeEach(() => {
       model.url = () => '/api'
       spy = jest.spyOn(apiClient(), 'post')
-      request = model.rpc('searching', 'search', {
-        method: 'GET'
-      })
+      request = model.rpc('search', { method: 'GET' }, 'searching')
     })
 
     it('returns a Request promise', () => {
@@ -134,17 +145,27 @@ describe(Base, () => {
     })
 
     it('sends a request using the endpoint suffix', () => {
-      expect(spy.mock.calls[0][0]).toBe('/api/search')
+      expect(spy.mock.calls.pop()[0]).toBe('/api/search')
     })
 
     it('passes the options to the api adapter', () => {
-      expect(spy.mock.calls[0][1]).toEqual({
+      expect(spy.mock.calls.pop()[1]).toEqual({
         method: 'GET'
       })
     })
 
     it('tracks the request with the specified labels', () => {
       expect(model.isRequest('searching')).toBe(true)
+    })
+
+    describe('rpc with rootUrl', () => {
+      beforeEach(() => {
+        model.rpc({ rootUrl: '/another_api/search' }, { method: 'GET' }, 'searching')
+      })
+
+      it('should fetch with the rootUrl', () => {
+        expect(spy.mock.calls.pop()[0]).toBe('/another_api/search')
+      })
     })
   })
 })