+++ __tests__/utils.ts
@@ -1,0 +1,7 @@
+export function strMapToObj(strMap) {
+    let obj = Object.create(null)
+
+    strMap.forEach((val, key) => obj[key] = val)
+
+    return obj
+}