+++ src/ErrorObject.ts
@@ -1,9 +1,9 @@
-// @flow
 export default class ErrorObject {
-  requestResponse: mixed
-  error: mixed
+  error: any = null
+  payload: any = {}
+  requestResponse: any = null
 
-  constructor (error: { requestResponse: mixed, error: mixed } | string | Error) {
+  constructor (error: { requestResponse: any, error: any } | string | Error) {
     if (error instanceof Error) {
       console.error(error)
       this.requestResponse = null
@@ -11,9 +11,11 @@ export default class ErrorObject {
     } else if (typeof error === 'string') {
       this.requestResponse = null
       this.error = error
-    } else {
+    } else if (error.requestResponse || error.error ){
       this.requestResponse = error.requestResponse
       this.error = error.error
+    } else {
+      this.payload = error
     }
   }
 }